improvement: Check if app exists before trying to delete it

+++ controllers/service.js
@@ -166,7 +166,11 @@ async function deleteApp(req, res) {
       message: "Missing parameters. [app_id]",
     })
 
-  const { status: appDeleteStatus, data } = appDb.deleteApp(appId)
+  const exists = await getAppInformation(appId)
+  if(!exists)
+    return res.status(404).json({
+      message: "App doesn't exist"
+    })
   const { status: appDeleteStatus, data } = await appDb.deleteApp(appId)
   if (appDeleteStatus !== "success")
     return res.status(500).json({
